+++ src/controllers/comment.php
@@ -2,8 +2,8 @@
 
 namespace App\Controllers;
 
-require_once('src/lib/database.php');
-require_once('src/model/comment.php');
+// require_once('src/lib/database.php');
+// require_once('src/model/comment.php');
 
 use App\Lib\Database;
 use App\Model\CommentRepository;
