+++ src/controllers/homepage.php
@@ -2,8 +2,8 @@
 
 namespace App\Controllers;
 
-require_once('src/lib/database.php');
-require_once('src/model/post.php');
+// require_once('src/lib/database.php');
+// require_once('src/model/post.php');
 
 use App\Lib\Database;
 use App\Model\PostRepository;
