+++ src/controllers/post.php
@@ -2,9 +2,9 @@
 
 namespace App\Controllers;
 
-require_once('src/lib/database.php');
-require_once('src/model/comment.php');
-require_once('src/model/post.php');
+// require_once('src/lib/database.php');
+// require_once('src/model/comment.php');
+// require_once('src/model/post.php');
 
 // spl_autoload_register(function($fqcn){
 // 	$path = str_replace(['App', '\\'], ['src', '/'], $fqcn) . '.php';
