+++ src/model/post.php
@@ -2,13 +2,12 @@
 declare(strict_types=1);
 namespace App\Model;
 
-require_once('src/lib/database.php');
-require_once('src/model/traits/idTrait.php');
-require_once('src/model/traits/titleTrait.php');
-require_once('src/model/traits/contentTrait.php');
-require_once('src/model/traits/createdAtTrait.php');
+// require_once('src/lib/database.php');
+// require_once('src/model/traits/idTrait.php');
+// require_once('src/model/traits/titleTrait.php');
+// require_once('src/model/traits/contentTrait.php');
+// require_once('src/model/traits/createdAtTrait.php');
 
-use App\Lib\Database;
 use App\Model\Traits\contentTrait;
 use App\Model\Traits\createdAtTrait;
 use App\Model\Traits\idTrait;
@@ -59,40 +58,4 @@ class Post
 
 		return $this;
 	}
-}
-
-class PostRepository
-{
-    public Database $connection;
-
-    public function getPost(int $id): Post
-    {
-        $statement = $this->connection->getConnection()->prepare('SELECT * FROM post WHERE id = ?');
-        $statement->execute([$id]);
-
-        $row = $statement->fetch();
-        $post = new Post();
-        $post->setId($row['id']);
-        $post->setTitle($row['title']);
-        $post->setContent($row['content']);
-
-        return $post;
-    }
-
-    public function getPosts():array
-    {
-        $statement = $this->connection->getConnection()->query('SELECT * FROM post');
-
-        $posts = [];
-        while (($row = $statement->fetch())) {
-            $post = new Post();
-            $post->setId($row['id']);
-            $post->setTitle($row['title']);
-            $post->setContent($row['content']);
-    
-            $posts[] = $post;
-        }
-    
-        return $posts;
-    }
 }
